refactor(VList): move around stuff to reduce circular deps
