+++ packages/vuetify/src/components/VList/VListItem.tsx
@@ -18,6 +18,7 @@ import { makeRouterProps, useLink } from '@/composables/router'
 import { makeTagProps } from '@/composables/tag'
 import { makeThemeProps, provideTheme } from '@/composables/theme'
 import { genOverlays, makeVariantProps, useVariant } from '@/composables/variant'
+import { useList } from './list'
 
 // Directives
 import { Ripple } from '@/directives/ripple'
@@ -26,7 +27,6 @@ import { Ripple } from '@/directives/ripple'
 import { computed, onMounted } from 'vue'
 import { genericComponent } from '@/util'
 import { useNestedItem } from '@/composables/nested/nested'
-import { useList } from './VList'
 
 // Types
 import type { MakeSlots } from '@/util'
