+++ packages/vuetify/src/components/VList/list.ts
@@ -1,0 +1,42 @@
+import { computed, inject, provide, ref } from 'vue'
+
+// Types
+import type { InjectionKey, Ref } from 'vue'
+
+// Depth
+export const DepthKey: InjectionKey<Ref<number>> = Symbol.for('vuetify:depth')
+
+export const useDepth = (hasPrepend?: Ref<boolean>) => {
+  const parent = inject(DepthKey, ref(-1))
+
+  const depth = computed(() => parent.value + 1 + (hasPrepend?.value ? 1 : 0))
+
+  provide(DepthKey, depth)
+
+  return depth
+}
+
+// List
+export const ListKey: InjectionKey<{
+  hasPrepend: Ref<boolean>
+  updateHasPrepend: (value: boolean) => void
+}> = Symbol.for('vuetify:list')
+
+export const createList = () => {
+  const parent = inject(ListKey, { hasPrepend: ref(false), updateHasPrepend: () => null })
+
+  const data = {
+    hasPrepend: ref(false),
+    updateHasPrepend: (value: boolean) => {
+      if (value) data.hasPrepend.value = value
+    },
+  }
+
+  provide(ListKey, data)
+
+  return parent
+}
+
+export const useList = () => {
+  return inject(ListKey, null)
+}
